refactor: move logic from app to main page component
