+++ src/pages/main/MainPage.tsx
@@ -1,0 +1,54 @@
+import { useEffect, useState } from 'react';
+import Search from '../../components/Search/Search';
+import SearchResults from '../../components/SearchResults/SearchResults';
+import { getResponse } from '../../api/api';
+
+type AppState = {
+  isLoading: boolean;
+  isError: boolean;
+  searchString: string;
+  searchResults: unknown[];
+};
+
+function MainPage() {
+  const [appState, setAppState] = useState<AppState>({
+    isLoading: false,
+    isError: false,
+    searchString: localStorage.getItem('searchString') ?? '',
+    searchResults: [],
+  });
+
+  async function onSearch(searchString: string) {
+    try {
+      setAppState({ ...appState, isLoading: true });
+      localStorage.setItem('searchString', searchString);
+      const response = await getResponse(searchString);
+      const results = await response.json();
+
+      if (response.ok) {
+        setAppState({ ...appState, searchResults: results, isLoading: false });
+      } else {
+        setAppState({ ...appState, isLoading: false, isError: true });
+      }
+    } catch {
+      setAppState({ ...appState, isLoading: false, isError: true });
+    }
+  }
+
+  useEffect(() => {
+    onSearch(appState.searchString);
+  }, [appState.searchString]);
+
+  return (
+    <div className="wrapper">
+      <Search onSearch={(searchString) => onSearch(searchString)} />
+      <SearchResults
+        isError={appState.isError}
+        isLoading={appState.isLoading}
+        searchResults={appState.searchResults}
+      />
+    </div>
+  );
+}
+
+export default MainPage;
